multi agent

+++ src/attack_simulator/ids.py
@@ -9,7 +9,7 @@ from attack_simulator.constants import UINT
 
 # TODO: Move functionality from simulator to this class.
 class IDS:
-    def __init__(self, seed: UINT, fnr: float, fpr: float) -> None:
+    def __init__(self, seed: int, fnr: float, fpr: float) -> None:
         self.seed = seed
         self.fnr = fnr
         self.fpr = fpr
@@ -19,10 +19,10 @@ class ProbabilityIDS(IDS):
     def ids_function(self, obs: UINT, p: float) -> UINT:
         if obs == 1:
             if p <= self.fnr:
-                return not obs
+                return UINT(not obs)
             return obs
         if p <= self.fpr:
-            return not obs
+            return UINT(not obs)
         return obs
 
     def __call__(
